Adding test project.

+++ CommifyTaxCalculatorAPI/Models/TaxBand.cs
@@ -4,7 +4,7 @@ public class TaxBand
 {
     public int TaxBandId { get; set; }
     public string TaxBandName { get; set; }
-    public decimal TaxBandRangeStart { get; set; }
-    public decimal TaxBandRangeEnd { get; set; }
-    public double TaxBandRate { get; set; }
+    public int TaxBandRangeStart { get; set; }
+    public int TaxBandRangeEnd { get; set; }
+    public decimal TaxBandRate { get; set; }
 }
